remove more theme usage and move wallets folder to store
