+++ components/ui/Checkbox.tsx
@@ -8,7 +8,7 @@ import { useUserLocation } from "lib/hooks/useUserLocation";
 import { observer } from "mobx-react";
 import { ChangeEvent } from "react";
 
-const useStyles = makeStyles<Theme, { theme: string }>({
+const useStyles = makeStyles({
   root: {
     "&:hover": {
       backgroundColor: "transparent",
@@ -18,7 +18,7 @@ const useStyles = makeStyles<Theme, { theme: string }>({
     borderRadius: 3,
     width: 16,
     height: 16,
-    backgroundColor: (props) => (props.theme === "dark" ? "black" : "white"),
+    backgroundColor: (props) => "white",
   },
   checkedIcon: {
     backgroundImage:
@@ -45,7 +45,7 @@ interface CheckboxProps {
 const Checkbox = observer(({ value, onChange, disabled }: CheckboxProps) => {
   const userStore = useUserLocation();
 
-  const classes = useStyles({ theme: userStore.theme });
+  const classes = useStyles();
 
   return (
     <MaterialCheckBox
