+++ lib/stores/wallets/errors/AuthError.tsx
