+++ lib/stores/wallets/errors/BaseWalletError.tsx
